requirements.txt
streamlit==1.34.0
pandas
requests
openai
pinecone-client
langchain
python-dotenv